+++ docs/changelog.rst
@@ -14,6 +14,7 @@ Changelog
 * :feature:`6033` Fix gas fee calculation for Optimism transactions to include L1 fees.
 * :bug:`-` ENS names that use the new RegistrarController and are renewed will have their events properly detected.
 * :bug:`-` Improve date and hexadecimal address scrambling.
+* :feature:`-` Transactions for bridging assets between Arbitrum One and Ethereum via the arbitrum bridge will now be decoded properly.
 
 * :release:`1.29.1 <2023-07-27>`
 * :bug:`-` Fix a bug where some images indicating the location of the assets are not loaded.